Auth test fixes, lease tests for individual opportunities, customer absent/attended fix, and broker microservice logging fixes
