+++ packages/openactive-integration-tests/package.json
@@ -18,7 +18,7 @@
   "eslintConfig": {},
   "dependencies": {
     "@openactive/data-model-validator": "^2.0.52",
-    "@openactive/data-models": "^2.0.234",
+    "@openactive/data-models": "^2.0.236",
     "@openactive/openactive-openid-test-client": "file:../openactive-openid-test-client",
     "@openactive/test-interface-criteria": "file:../test-interface-criteria",
     "async-mutex": "^0.3.1",
