+++ packages/openactive-integration-tests/test/features/README.md
@@ -41,11 +41,11 @@ The tests for these features cover all known edge cases, including both happy an
 | details-capture | Simple Book including Attendee Details capture ([attendee-details-capture](./details-capture/attendee-details-capture/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#attendee-details-capture) | Support for capturing attendee details | [TestOpportunityBookableAttendeeDetails](https://openactive.io/test-interface#TestOpportunityBookableAttendeeDetails) x6, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x2 |
 | details-capture | Customer Details identifier capture ([customer-details-capture-identifier](./details-capture/customer-details-capture-identifier/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#schema-person-for-customer-or-attendee) | Support for capturing the Broker's identifier for the customer. Note this field cannot be mandatory. | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
 | details-capture | Customer Details non-essential capture ([customer-details-capture-non-essential](./details-capture/customer-details-capture-non-essential/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#customer-details-capture) | Support for capturing forename, surname, and telephone number from the Customer. Note these fields cannot be mandatory. | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
-| leasing | Anonymous leasing, including leaseExpires ([anonymous-leasing](./leasing/anonymous-leasing/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#leasing) | Leasing at C1, reserving a space in the opportunity before the Customer has provided their contact information | [TestOpportunityBookableFiveSpaces](https://openactive.io/test-interface#TestOpportunityBookableFiveSpaces) x2, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
-| leasing | Named leasing, including leaseExpires ([named-leasing](./leasing/named-leasing/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#leasing) | Leasing at C2, reserving a space in the opportunity after the Customer has provided their contact information | [TestOpportunityBookableFiveSpaces](https://openactive.io/test-interface#TestOpportunityBookableFiveSpaces) x2, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
+| leasing | Anonymous leasing, including leaseExpires ([anonymous-leasing](./leasing/anonymous-leasing/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#leasing) | Leasing at C1, reserving a space in the opportunity before the Customer has provided their contact information | [TestOpportunityBookableFiveSpaces](https://openactive.io/test-interface#TestOpportunityBookableFiveSpaces) x2, [TestOpportunityBookableOneSpace](https://openactive.io/test-interface#TestOpportunityBookableOneSpace) x2, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
+| leasing | Named leasing, including leaseExpires ([named-leasing](./leasing/named-leasing/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#leasing) | Leasing at C2, reserving a space in the opportunity after the Customer has provided their contact information | [TestOpportunityBookableFiveSpaces](https://openactive.io/test-interface#TestOpportunityBookableFiveSpaces) x2, [TestOpportunityBookableOneSpace](https://openactive.io/test-interface#TestOpportunityBookableOneSpace) x2, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
 | notifications | Change of logistics notifications ([change-of-logistics-notifications](./notifications/change-of-logistics-notifications/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#change-of-logistics-notifications) | Notifications for when an opportunity's name, location, or start/end date/time are updated  | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
 | notifications | Customer notice notifications ([customer-notice-notifications](./notifications/customer-notice-notifications/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#customer-notice-notifications) | Text notifications broadcast to all registered attendees of an opportunity | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
-| notifications | Opportunity attendance updates ([opportunity-attendance-updates](./notifications/opportunity-attendance-updates/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#opportunity-attendance-updates) | Allowing the broker to recieve updates for when an attendee attends an event | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x4 |
+| notifications | Opportunity attendance updates ([opportunity-attendance-updates](./notifications/opportunity-attendance-updates/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#opportunity-attendance-updates) | Allowing the broker to recieve updates for when an attendee attends an event | [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x8 |
 | payment | Free opportunities ([free-opportunities](./payment/free-opportunities/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#free-opportunities) | The most simple form of booking, for free opportunities. Does not check for leases. | [TestOpportunityBookableFree](https://openactive.io/test-interface#TestOpportunityBookableFree) x10 |
 | payment | Opportunities with a non-zero price ([non-free-opportunities](./payment/non-free-opportunities/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#step-by-step-process-description) | The most simple form of booking with payment. Does not check for leases. | [TestOpportunityBookableNonFree](https://openactive.io/test-interface#TestOpportunityBookableNonFree) x6, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x2 |
 | payment | Payment reconciliation detail validation ([payment-reconciliation-detail-validation](./payment/payment-reconciliation-detail-validation/README.md)) | Optional<br>[View Spec](https://www.openactive.io/open-booking-api/EditorsDraft/#payment-reconciliation-detail-validation) | Booking with valid, invalid, and missing Payment details | [TestOpportunityBookableFree](https://openactive.io/test-interface#TestOpportunityBookableFree) x12, [TestOpportunityBookableUsingPayment](https://openactive.io/test-interface#TestOpportunityBookableUsingPayment) x15, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x5 |
