+++ packages/openactive-integration-tests/test/features/authentication/booking-partner-authentication/implemented/authorization-code-flow-test.js
@@ -27,7 +27,7 @@ function (configuration, orderItemCriteria, featureIsImplemented, logger) {
       .setClientCredentials(false, 'primary')
       .authorizeAuthorizationCodeFlow({
         loginCredentialsAccessor: () => SELLER_CONFIG.primary.authentication.loginCredentials,
-        assertFlowRequiredConsent: true,
+        assertFlowRequiredConsent: false, // TODO: Add a test interface Action that resets consent for the specified user, and call it before this flow starts. Then this flow should assert consent.
         assertSellerIdClaim: SELLER_CONFIG.primary['@id'],
       })
       .refresh();
