+++ packages/openactive-integration-tests/test/features/cancellation/customer-requested-cancellation/README.md
@@ -37,7 +37,7 @@ Update `default.json` within `packages/openactive-integration-tests/config/` as 
 
 | Identifier | Name | Description | Prerequisites per Opportunity Type |
 |------------|------|-------------|---------------|
-| [atomic-cancel](./implemented/atomic-cancel-test.js) | Successful booking and successful cancellation after atomic failed cancellation request | After a successful booking, and an unsuccessful but atomic cancellation request, successfully cancel, including checking the Orders feed. | [TestOpportunityBookableCancellable](https://openactive.io/test-interface#TestOpportunityBookableCancellable) x2, [TestOpportunityBookableNotCancellable](https://openactive.io/test-interface#TestOpportunityBookableNotCancellable) x1 |
+| [atomic-cancel](./implemented/atomic-cancel-test.js) | Successful booking and successful cancellation after atomic failed cancellation request | After a successful booking, and an unsuccessful but atomic cancellation request, successfully cancel, including checking the Orders feed. Atomic means that the cancellation request either entirely succeeds (all OrderItems items are cancelled) or entirely fails (no OrderItems are cancelled), which is generally best achieved by wrapping code in a database transaction. | [TestOpportunityBookableCancellable](https://openactive.io/test-interface#TestOpportunityBookableCancellable) x2, [TestOpportunityBookableNotCancellable](https://openactive.io/test-interface#TestOpportunityBookableNotCancellable) x1 |
 | [book-and-cancel](./implemented/book-and-cancel-test.js) | Successful booking and cancellation. | A successful end to end booking including full Order cancellation, including checking the Orders Feed. Two cancellation requests are made to ensure that cancellation is atomic. | [TestOpportunityBookableCancellable](https://openactive.io/test-interface#TestOpportunityBookableCancellable) x4 |
 | [orderitem-id-invalid-error](./implemented/orderitem-id-invalid-error-test.js) | Expect a OrderItemIdInvalidError for an Order that does not exist | Runs Order Cancellation for a non-existent invalid OrderItem, but real Order, expecting an OrderItemIdInvalidError error to be returned | [TestOpportunityBookableCancellable](https://openactive.io/test-interface#TestOpportunityBookableCancellable) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
 | [orderitem-not-within-order-error](./implemented/orderitem-not-within-order-error-test.js) | Expect a OrderItemNotWithinOrderError for an Order that does not exist | Runs Order Cancellation for a non-existent Order (with a fictional UUID), but real OrderItem, expecting an OrderItemNotWithinOrderError error to be returned | [TestOpportunityBookableCancellable](https://openactive.io/test-interface#TestOpportunityBookableCancellable) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
