+++ packages/openactive-integration-tests/test/features/notifications/opportunity-attendance-updates/implemented/opportunity-attendance-update-customer-absent-test.js
@@ -6,17 +6,17 @@ FeatureHelper.describeFeature(module, {
   testCategory: 'notifications',
   testFeature: 'opportunity-attendance-updates',
   testFeatureImplemented: true,
-  testIdentifier: 'opportunity-attendance-updates',
-  testName: "Changes to an opportunity's attendance (via OpportunityAttendanceUpdateSimulateAction) should update the Order Feed.",
-  testDescription: "After B, invoke an OpportunityAttendanceUpdateSimulateAction. This should create an update in the Order Feed with the OrderItem's orderItemStatus changed to CustomerAttended",
+  testIdentifier: 'opportunity-attendance-update-customer-absent',
+  testName: "Changes to an opportunity's attendance (via AttendeeAbsentSimulateAction) should update the Order Feed.",
+  testDescription: "After B, invoke an `AttendeeAbsentSimulateAction`. This should create an update in the Order Feed with the OrderItem's orderItemStatus changed to `https://openactive.io/CustomerAttended`",
   // The primary opportunity criteria to use for the primary OrderItem under test
   testOpportunityCriteria: 'TestOpportunityBookable',
   controlOpportunityCriteria: 'TestOpportunityBookable',
 },
-TestRecipes.simulateActionAndExpectOrderFeedUpdateAfterSimpleC1C2Book({ actionType: 'test:OpportunityAttendanceUpdateSimulateAction' },
+TestRecipes.simulateActionAndExpectOrderFeedUpdateAfterSimpleC1C2Book({ actionType: 'test:AttendeeAbsentSimulateAction' },
   ({ orderFeedUpdate, orderItemCriteriaList }) => {
     it('OrderItems should have CustomerAttended statuses', () => {
-      // new = after the OpportunityAttendanceUpdateSimulateAction was invoked
+      // new = after the AttendeeAbsentSimulateAction was invoked
       const newOrderItems = orderFeedUpdate.getOutput().httpResponse.body.data.orderedItem;
       // As we'll be setting out expectations in an iteration, this test would
       // give a false positive if there were no items in `orderedItem`, so we
@@ -29,7 +29,7 @@ TestRecipes.simulateActionAndExpectOrderFeedUpdateAfterSimpleC1C2Book({ actionTy
         const { orderItemStatus } = newOrderItem;
 
         expect(orderItemStatus).to.be.a('string');
-        expect(orderItemStatus).to.be.equal('https://openactive.io/CustomerAttended');
+        expect(orderItemStatus).to.be.equal('https://openactive.io/AttendeeAbsent');
       }
     });
   }));
