+++ packages/openactive-integration-tests/test/helpers/flow-stages/test-interface-action.js
@@ -15,7 +15,8 @@ const { FlowStageUtils } = require('./flow-stage-utils');
  *   | 'test:ChangeOfLogisticsNameSimulateAction'
  *   | 'test:ChangeOfLogisticsLocationSimulateAction'
  *   | 'test:CustomerNoticeSimulateAction'
- *   | 'test:OpportunityAttendanceUpdateSimulateAction'
+ *   | 'test:AttendeeAttendedSimulateAction'
+ *   | 'test:AttendeeAbsentSimulateAction'
  *   | 'test:ReplacementSimulateAction'
  *   | 'test:SellerAcceptOrderProposalSimulateAction'
  *   | 'test:SellerAmendOrderProposalSimulateAction'
