+++ packages/test-interface-criteria/src/criteria/TestOpportunityBookableAdditionalDetails.js
@@ -45,7 +45,7 @@ const TestOpportunityBookableAdditionalDetails = createCriteria({
   ],
   testDataShape: (options) => ({
     opportunityConstraints: {
-      ...shapeConstraintRecipes.remainingCapacityMustBeAtLeastTwo(),
+      ...shapeConstraintRecipes.remainingCapacityMustBeAtLeast(2),
       ...shapeConstraintRecipes.sellerMustAllowOpenBooking(),
     },
     offerConstraints: {
